wtvdec: fix integer overflow
Fixes CID968583

+++ libavformat/wtvdec.c
@@ -200,7 +200,7 @@ static AVIOContext * wtvfile_open_sector(int first_sector, uint64_t length, int 
         return NULL;
     }
 
-    if (wf->sectors[wf->nb_sectors - 1] << WTV_SECTOR_BITS > avio_tell(s->pb))
+    if ((int64_t)wf->sectors[wf->nb_sectors - 1] << WTV_SECTOR_BITS > avio_tell(s->pb))
         av_log(s, AV_LOG_WARNING, "truncated file\n");
 
     /* check length */
